--- app/footer.css
@@ -1,73 +1,0 @@
-.footer {
-    background-color: #3498db;
-    color: white;
-    padding: 40px 20px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .footer-content {
-    width: 100%;
-    max-width: 1200px;
-    display: flex;
-    justify-content: space-between;
-  }
-  
-  .footer-section {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .footer-logo {
-    margin-bottom: 20px;
-  }
-  
-  .footer-logo img {
-    width: 50px; /* Adjust as needed */
-  }
-  
-  .footer-logo p {
-    max-width: 300px;
-    margin-top: 10px;
-    color: white;
-  }
-  
-  .footer-section h3 {
-    margin-bottom: 15px;
-  }
-  
-  .footer-section ul {
-    list-style: none;
-    padding: 0;
-  }
-  
-  .footer-section ul li a {
-    color: white;
-    text-decoration: none;
-    transition: color 0.3s;
-  }
-  
-  .footer-section ul li a:hover {
-    color: #fff;
-  }
-  
-  .social-icons a {
-    display: inline-block;
-    margin-right: 15px;
-    color: #aaa;
-  }
-  
-  .social-icons a:hover {
-    color: #fff;
-  }
-  
-  .footer-bottom {
-    margin-top: 20px;
-    text-align: center;
-    font-size: 0.8rem;
-    color: #aaa;
-  }
-  
-  /* Add media queries for responsiveness as necessary */
-
