+++ package.json
@@ -10,17 +10,13 @@
   },
   "dependencies": {
     "bcrypt": "^5.1.1",
-    "jquery": "^3.7.1",
     "mongodb": "^6.5.0",
     "mongoose": "^8.2.2",
     "next": "14.1.3",
     "next-auth": "^4.24.7",
     "pg": "^8.11.3",
     "react": "^18",
-    "react-dom": "^18",
-    "react-icons": "^5.1.0",
-    "react-slick": "^0.30.2",
-    "slick-carousel": "^1.8.1"
+    "react-dom": "^18"
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
